// apps/desktop/frontend/bindings/surveycontroller/surveycore/internal/model/models.ts
export enum QuestionKind {
    /**
     * The Go zero value for the underlying type of the enum.
     */
    $zero = "",

    QuestionKindSingle = "single",
    QuestionKindMultiple = "multiple",
    QuestionKindDropdown = "dropdown",
    QuestionKindScale = "scale",
    QuestionKindMatrix = "matrix",
    QuestionKindOrder = "order",
    QuestionKindSlider = "slider",
    QuestionKindText = "text",
    QuestionKindMultiText = "multi_text",
};